Moved mapping logic to its own file.
